Fixed the signed value for start position and start record in BACnet file object abstraction. (#1057)

+++ ports/posix/bacfile-posix.h
@@ -21,26 +21,26 @@ bool bacfile_posix_file_size_set(const char *pathname, size_t file_size);
 BACNET_STACK_EXPORT
 size_t bacfile_posix_read_stream_data(
     const char *pathname,
-    size_t fileStartPosition,
+    int32_t fileStartPosition,
     uint8_t *fileData,
     size_t fileDataLen);
 BACNET_STACK_EXPORT
 size_t bacfile_posix_write_stream_data(
     const char *pathname,
-    size_t fileStartPosition,
+    int32_t fileStartPosition,
     const uint8_t *fileData,
     size_t fileDataLen);
 BACNET_STACK_EXPORT
 bool bacfile_posix_write_record_data(
     const char *pathname,
-    size_t fileStartRecord,
+    int32_t fileStartRecord,
     size_t fileIndexRecord,
     const uint8_t *fileData,
     size_t fileDataLen);
 BACNET_STACK_EXPORT
 bool bacfile_posix_read_record_data(
     const char *pathname,
-    size_t fileStartRecord,
+    int32_t fileStartRecord,
     size_t fileIndexRecord,
     uint8_t *fileData,
     size_t fileDataLen);
